src/requirements.txt
pyyaml
streamlit
fastapi
uvicorn
python-multipart
requests